Add click cli framework
Use click framework to have easier CLI support.
One single script to support all functionality instead of having several
scripts.
The single script have several commands to encapsulate the functionality
including --help section and description.
