+++ requirements.txt
@@ -1,2 +1,3 @@
 BeautifulSoup==3.2.1
+click==5.1
 wsgiref==0.1.2
