+++ scraper.py
@@ -1,0 +1,5 @@
+from image_scraper.cli.commands import cli
+
+
+if __name__ == "__main__":
+    cli()
